Add files via upload
If you want to run everything all at once, download this code and put it in the same file as all the other scripts. 

Be aware, the amico and dasilva files have the same name, however they are different from the scripts in the "code" folder and should be replaced with these using the Loopcode.